refactor(autoware_qp_interface): keep status_to_string internal-linkage and fold status tests
Address review: move status_to_string() into an anonymous namespace in
osqp_interface.cpp (no exported symbol), drop the internal
src/osqp_interface_status.hpp and its PRIVATE include dir, remove the
change-detector StatusToString test, and fold the two behavioral tests
(InfeasibleProblemReportsNotSolved, FreshlyConstructedIsNotSolved) into
test_osqp_interface.cpp. The bug fix (getStatus() reporting the real solver
status) stays covered by behavioral tests through the public API.
